+++ example/lib/enums/post_action.dart
@@ -1,0 +1,1 @@
+enum PostAction { delete, markRead }